게시글 작성 추가

+++ src/main/webapp/jsp/article/detail.jsp
@@ -31,7 +31,9 @@
 	</div>
 	<div>
 		<a href="list">목록</a>
-		<a href="delete?id=<%= articleMap.get("id") %>">삭제</a>
+	</div>
+	<div>
+		<a href="delete?id=<%= articleMap.get("id") %>" onclick = "if(confirm('정말 삭제하시겠습니까?') == false) return false;" >삭제</a>
 	</div>
 
 </body>
